Remove option parsing entirely
I don't want to interfere in the original paleta's version namespace if
it so happens that Dylan Araps returns. So, we'll just not return a
version, instead. The version should be of little importance since the
format is so simple, and anyone installing it should have access to a
package version anyway.